moved test components in Test folder
