+++ src/components/Test/Speaking/Speaking.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
-import { offRecAudio, onRecAudio } from "../../services/voice-recorder";
-import { startTimer, createTimer, stopTimer } from "../../services/timer";
+import { offRecAudio, onRecAudio } from "../../../services/voice-recorder";
+import { startTimer, createTimer, stopTimer } from "../../../services/timer";
 import MicOffIcon from "@material-ui/icons/MicOff";
 import MicIcon from "@material-ui/icons/Mic";
-import { Player } from "../";
+import { Player } from "../../index";
 import "./Speaking.scss";
 
 export const Speaking = () => {
