+++ src/components/Test/testData/GrammarTasks.jsx
@@ -1,4 +1,4 @@
-const listeningTasks = [
+const grammarTasks = [
   {
     sentence: 'The best berries for me … strawberries.',
     options: [
@@ -82,7 +82,8 @@ const listeningTasks = [
         isRight: false
       }
     ]
-  },{
+  },
+  {
     sentence: 'The best berries for me … strawberries.',
     options: [
       {
@@ -123,7 +124,8 @@ const listeningTasks = [
         isRight: false
       }
     ]
-  },{
+  },
+  {
     sentence: 'The best berries for me … strawberries.',
     options: [
       {
@@ -164,7 +166,50 @@ const listeningTasks = [
         isRight: false
       }
     ]
-  },{
+  },
+  {
+    sentence: 'The best berries for me … strawberries.',
+    options: [
+      {
+        option: 'is',
+        isRight: false
+      },
+      {
+        option: 'are',
+        isRight: true
+      },
+      {
+        option: 'be',
+        isRight: false
+      },
+      {
+        option: 'am',
+        isRight: false
+      }
+    ]
+  },
+  {
+    sentence: 'He … swim very well. He is the best swimmer in the team!',
+    options: [
+      {
+        option: 'can',
+        isRight: true
+      },
+      {
+        option: 'must',
+        isRight: false
+      },
+      {
+        option: 'should',
+        isRight: false
+      },
+      {
+        option: 'have to',
+        isRight: false
+      }
+    ]
+  },
+  {
     sentence: 'The best berries for me … strawberries.',
     options: [
       {
@@ -208,4 +253,4 @@ const listeningTasks = [
   }
 ];
 
-export default listeningTasks;
+export {grammarTasks};
