+++ src/components/Test/testData/ListeningTasks.jsx
@@ -1,4 +1,4 @@
-const grammarTasks = [
+const listeningTasks = [
   {
     sentence: 'The best berries for me … strawberries.',
     options: [
@@ -82,8 +82,7 @@ const grammarTasks = [
         isRight: false
       }
     ]
-  },
-  {
+  },{
     sentence: 'The best berries for me … strawberries.',
     options: [
       {
@@ -124,8 +123,7 @@ const grammarTasks = [
         isRight: false
       }
     ]
-  },
-  {
+  },{
     sentence: 'The best berries for me … strawberries.',
     options: [
       {
@@ -166,50 +164,7 @@ const grammarTasks = [
         isRight: false
       }
     ]
-  },
-  {
-    sentence: 'The best berries for me … strawberries.',
-    options: [
-      {
-        option: 'is',
-        isRight: false
-      },
-      {
-        option: 'are',
-        isRight: true
-      },
-      {
-        option: 'be',
-        isRight: false
-      },
-      {
-        option: 'am',
-        isRight: false
-      }
-    ]
-  },
-  {
-    sentence: 'He … swim very well. He is the best swimmer in the team!',
-    options: [
-      {
-        option: 'can',
-        isRight: true
-      },
-      {
-        option: 'must',
-        isRight: false
-      },
-      {
-        option: 'should',
-        isRight: false
-      },
-      {
-        option: 'have to',
-        isRight: false
-      }
-    ]
-  },
-  {
+  },{
     sentence: 'The best berries for me … strawberries.',
     options: [
       {
@@ -253,4 +208,4 @@ const grammarTasks = [
   }
 ];
 
-export default grammarTasks;
+export {listeningTasks};
