+++ src/components/index.js
@@ -1,8 +1,10 @@
 export { EditTestsFilter } from './EditTestsFilter/EditTestsFilter';
 export { EditTestsTable } from './EditTestsTable/EditTestsTable';
 export { TestLevelSelector } from './TestLevelSelector/TestLevelSelector';
-export { Essay } from './Essay/Essay';
-export { Grammar } from './Grammar/Grammar';
-export { Listening } from './Listening/Listening';
-export { Speaking } from './Speaking/Speaking';
-export { Player } from './Player/Player';
+export { Essay } from './Test/Essay/Essay';
+export { Grammar } from './Test/Grammar/Grammar';
+export { Listening } from './Test/Listening/Listening';
+export { Speaking } from './Test/Speaking/Speaking';
+export { Player } from './Test/Player/Player';
+export { listeningTasks } from './Test/testData/ListeningTasks';
+export { grammarTasks } from './Test/testData/GrammarTasks';
